move string16 gin conv to new file
